requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
groq==0.9.0
python-docx==1.0.1
qrcode==7.4.2
reportlab==4.0.4
Pillow==10.0.1
python-dotenv==1.0.0